Split main away from NameGame, still have a few issues to be solved before it can run
